src/requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
art==6.5
boto3==1.39.8
botocore==1.39.8
certifi==2025.7.14
cffi==1.17.1
click==8.2.1
jmespath==1.0.1
minio==7.2.15
numpy==2.3.1
pandas==2.3.1
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.23.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
s3transfer==0.13.0
six==1.17.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
